Reduced unneeded net classes
