+++ gdx/src/com/badlogic/gdx.gwt.xml
@@ -315,6 +315,8 @@
 		<include name="net/ServerSocketHints.java"/>
 		<include name="net/Socket.java"/>
 		<include name="net/SocketHints.java"/>
+		<exclude name="net/NetJavaSocketImpl.java"/> <!-- Reason: Networking -->
+		<exclude name="net/NetJavaServerSocketImpl.java"/> <!-- Reason: Networking -->
 	
 	<!-- physics/box2d -->
 	<!-- Box2d is fully emulated in GWT backend -->
