+++ gdx/src/com/badlogic/gdx/net/NetJavaServerSocketImpl.java
@@ -13,7 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  ******************************************************************************/
-package com.badlogic.gdx.backends.lwjgl;
+
+package com.badlogic.gdx.net;
 
 import java.io.InputStream;
 import java.io.OutputStream;
@@ -26,45 +27,39 @@ import com.badlogic.gdx.net.Socket;
 import com.badlogic.gdx.net.SocketHints;
 import com.badlogic.gdx.utils.GdxRuntimeException;
 
-/**
- * Server socket implementation using java.net.ServerSocket.
+/** Server socket implementation using java.net.ServerSocket.
  * 
- * @author noblemaster
- */
-public class LwjglServerSocket implements ServerSocket {
+ * @author noblemaster */
+public class NetJavaServerSocketImpl implements ServerSocket {
 
 	private Protocol protocol;
-	
+
 	/** Our server or null for disposed, aka closed. */
 	private java.net.ServerSocket server;
-	
-	
-	public LwjglServerSocket(Protocol protocol, int port, ServerSocketHints hints) {
+
+	public NetJavaServerSocketImpl (Protocol protocol, int port, ServerSocketHints hints) {
 		this.protocol = protocol;
-		
+
 		// create the server socket
 		try {
 			// initialize
 			server = new java.net.ServerSocket();
 			if (hints != null) {
-				server.setPerformancePreferences(hints.performancePrefConnectionTime, 
-															hints.performancePrefLatency, 
-															hints.performancePrefBandwidth);
+				server.setPerformancePreferences(hints.performancePrefConnectionTime, hints.performancePrefLatency,
+					hints.performancePrefBandwidth);
 				server.setReuseAddress(hints.reuseAddress);
 				server.setSoTimeout(hints.acceptTimeout);
 				server.setReceiveBufferSize(hints.receiveBufferSize);
 			}
-			
+
 			// and bind the server...
 			InetSocketAddress address = new InetSocketAddress(port);
 			if (hints != null) {
 				server.bind(address, hints.backlog);
-			}
-			else {
+			} else {
 				server.bind(address);
 			}
-		}
-		catch (Exception e) {
+		} catch (Exception e) {
 			throw new GdxRuntimeException("Cannot create a server socket at port " + port + ".", e);
 		}
 	}
@@ -75,23 +70,21 @@ public class LwjglServerSocket implements ServerSocket {
 	}
 
 	@Override
-	public Socket accept(SocketHints hints) {
+	public Socket accept (SocketHints hints) {
 		try {
-			return new LwjglSocket(server.accept(), hints);
-		}
-		catch (Exception e) {
+			return new NetJavaSocketImpl(server.accept(), hints);
+		} catch (Exception e) {
 			throw new GdxRuntimeException("Error accepting socket.", e);
 		}
 	}
 
 	@Override
-	public void dispose() {
+	public void dispose () {
 		if (server != null) {
 			try {
 				server.close();
 				server = null;
-			}
-			catch (Exception e) {
+			} catch (Exception e) {
 				throw new GdxRuntimeException("Error closing server.", e);
 			}
 		}
