+++ gdx/src/com/badlogic/gdx/net/NetJavaSocketImpl.java
@@ -13,7 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  ******************************************************************************/
-package com.badlogic.gdx.backends.lwjgl;
+
+package com.badlogic.gdx.net;
 
 import java.io.InputStream;
 import java.io.OutputStream;
@@ -25,67 +26,59 @@ import com.badlogic.gdx.net.Socket;
 import com.badlogic.gdx.net.SocketHints;
 import com.badlogic.gdx.utils.GdxRuntimeException;
 
-/**
- * Socket implementation using java.net.Socket.
+/** Socket implementation using java.net.Socket.
  * 
- * @author noblemaster
- */
-public class LwjglSocket implements Socket {
+ * @author noblemaster */
+public class NetJavaSocketImpl implements Socket {
 
 	/** Our socket or null for disposed, aka closed. */
 	private java.net.Socket socket;
-	
-	
-	public LwjglSocket(Protocol protocol, String host, int port, SocketHints hints) {
+
+	public NetJavaSocketImpl (Protocol protocol, String host, int port, SocketHints hints) {
 		try {
 			// create the socket
 			socket = new java.net.Socket();
-			applyHints(hints);  // better to call BEFORE socket is connected!
-			
+			applyHints(hints); // better to call BEFORE socket is connected!
+
 			// and connect...
 			InetSocketAddress address = new InetSocketAddress(host, port);
 			if (hints != null) {
 				socket.connect(address, hints.connectTimeout);
-			}
-			else {
+			} else {
 				socket.connect(address);
 			}
-		}
-		catch (Exception e) {
+		} catch (Exception e) {
 			throw new GdxRuntimeException("Error making a socket connection to " + host + ":" + port, e);
 		}
 	}
-	
-	public LwjglSocket(java.net.Socket socket, SocketHints hints) {
+
+	public NetJavaSocketImpl (java.net.Socket socket, SocketHints hints) {
 		this.socket = socket;
 		applyHints(hints);
 	}
-	
-	private void applyHints(SocketHints hints) {
+
+	private void applyHints (SocketHints hints) {
 		if (hints != null) {
-			try {	
-				socket.setPerformancePreferences(hints.performancePrefConnectionTime, 
-					 										hints.performancePrefLatency, 
-					 										hints.performancePrefBandwidth);
+			try {
+				socket.setPerformancePreferences(hints.performancePrefConnectionTime, hints.performancePrefLatency,
+					hints.performancePrefBandwidth);
 				socket.setTrafficClass(hints.trafficClass);
 				socket.setTcpNoDelay(hints.tcpNoDelay);
 				socket.setKeepAlive(hints.keepAlive);
 				socket.setSendBufferSize(hints.sendBufferSize);
 				socket.setReceiveBufferSize(hints.receiveBufferSize);
 				socket.setSoLinger(hints.linger, hints.lingerDuration);
-			}
-			catch (Exception e) {
-				throw new GdxRuntimeException("Error setting socket hints." , e);
+			} catch (Exception e) {
+				throw new GdxRuntimeException("Error setting socket hints.", e);
 			}
 		}
 	}
-	
+
 	@Override
 	public boolean isConnected () {
 		if (socket != null) {
 			return socket.isConnected();
-		}
-		else {
+		} else {
 			return false;
 		}
 	}
@@ -94,8 +87,7 @@ public class LwjglSocket implements Socket {
 	public InputStream getInputStream () {
 		try {
 			return socket.getInputStream();
-		}
-		catch (Exception e) {
+		} catch (Exception e) {
 			throw new GdxRuntimeException("Error getting input stream from socket.", e);
 		}
 	}
@@ -104,27 +96,25 @@ public class LwjglSocket implements Socket {
 	public OutputStream getOutputStream () {
 		try {
 			return socket.getOutputStream();
-		}
-		catch (Exception e) {
+		} catch (Exception e) {
 			throw new GdxRuntimeException("Error getting output stream from socket.", e);
 		}
 	}
 
 	@Override
-	public void dispose() {
+	public String getRemoteAddress () {
+		return socket.getRemoteSocketAddress().toString();
+	}
+
+	@Override
+	public void dispose () {
 		if (socket != null) {
 			try {
 				socket.close();
 				socket = null;
-			}
-			catch (Exception e) {
+			} catch (Exception e) {
 				throw new GdxRuntimeException("Error closing socket.", e);
 			}
 		}
-	}
-	
-	@Override
-	public String getRemoteAddress () {
-		return socket.getRemoteSocketAddress().toString();
 	}
 }
